fixup! TF-2708 Allow Chrome to download PDF after view
