+++ core/lib/utils/html/js_interop_stub.dart
@@ -1,0 +1,5 @@
+import 'dart:typed_data';
+
+extension Uint8ListExtension on Uint8List {
+  dynamic get toJS => throw UnimplementedError();
+}
